[#54] order 주문 도메인 entity dto 비즈니스 로직 구현

+++ backend/src/main/java/com/example/pproject/user/dto/UserRequestDTO.java
@@ -17,7 +17,7 @@ public class UserRequestDTO {
     // ✅ member 테이블 구조(요구사항) 기준
     // - 기존 코드 호환을 위해 userid/username/birthday 필드는 유지
     // =========================
-    private Integer id;                 // (member_id) 일련번호
+    private Long id;                 // (member_id) 일련번호
 
     private String userid;              // (login_id) 아이디(일반 가입). 소셜 가입은 NULL 가능
     private String password;            // (password_hash) 비밀번호(요청에서만 사용)
